Add helper library for running Ruby scripts
This patch adds a puppetlabs.bodgery.jruby library that contains helpers
for running ruby scripts that are placed under `src/ruby` in the context of
Puppet Server's JRuby pool. As indicated by the name, this library is a
pile of hacks that shouldn't be let anywhere near a production instance.