Add Puppet Server as a development dependency
This patch adds Puppet Server 5.0.0 as a development dependency and submodule.
TrapperKeeper configuration files are copied over from the dependency. I'm not
the biggest fan of this since it is messy and brittle, but there isn't really
a better way to handle the configuration and provisioning of Ruby dependencies
unless Puppet Server starts exposing this info explicitly as part of a "dev"
classification.