Use java.io.File for listing directory contents
This patch switches from the NIO based Files/list back to using the .listFiles
method of the basic Java File class. The NIO method returns a DirectoryStream
which must be closed after iteration or else it will leak file descriptors.
For now, calling .listFiles is much simpler compared to figuring out how to
wrap with-open around the DirectoryStream.